added unit tests for useEvent
