+++ VirtualList/tests/VirtualGridList-specs.js
@@ -1,23 +1,59 @@
 import '@testing-library/jest-dom';
-import {act, render, screen} from '@testing-library/react';
+import {act, fireEvent, render, screen} from '@testing-library/react';
 
 import ImageItem from '../../ImageItem';
 import {VirtualGridList} from '../VirtualList';
 
-describe('VirtualGridList with native \'scrollMode\'', () => {
+describe('VirtualGridList with translate \'scrollMode\'', () => {
 	let
 		clientSize,
 		dataSize,
+		getScrollTo,
+		handlerOnScroll,
+		handlerOnScrollStart,
+		handlerOnScrollStop,
 		items,
 		itemSize,
+		myScrollTo,
+		onScrollCount,
+		onScrollStartCount,
+		onScrollStopCount,
 		renderItem,
+		resultScrollLeft,
+		resultScrollTop,
+		startScrollTop,
 		svgGenerator;
 
 	beforeEach(() => {
 		clientSize = {clientWidth: 1280, clientHeight: 720};
 		dataSize = 200;
 		items = [];
-		itemSize = {minWidth: 300, minHeight:240};
+		itemSize = {minWidth: 300, minHeight: 240};
+		onScrollCount = 0;
+		onScrollStartCount = 0;
+		onScrollStopCount = 0;
+		resultScrollLeft = 0;
+		resultScrollTop = 0;
+		startScrollTop = 0;
+
+		getScrollTo = (scrollTo) => {
+			myScrollTo = scrollTo;
+		};
+		handlerOnScroll = () => {
+			onScrollCount++;
+		};
+		handlerOnScrollStart = (e) => {
+			startScrollTop = e.scrollTop;
+			onScrollStartCount++;
+		};
+		handlerOnScrollStop = (done, testCase) => (e) => {
+			onScrollStopCount++;
+			resultScrollLeft = e.scrollLeft;
+			resultScrollTop = e.scrollTop;
+
+			testCase();
+			done();
+		};
 
 		renderItem = ({index, ...rest}) => { // eslint-disable-line enact/display-name
 			const {text, subText, source} = items[index];
@@ -52,9 +88,21 @@ describe('VirtualGridList with native \'scrollMode\'', () => {
 	afterEach(() => {
 		clientSize = null;
 		dataSize = null;
+		getScrollTo = null;
+		handlerOnScroll = null;
+		handlerOnScrollStart = null;
+		handlerOnScrollStop = null;
 		items = null;
 		itemSize = null;
+		myScrollTo = null;
+		onScrollCount = null;
+		onScrollStartCount = null;
+		onScrollStopCount = null;
 		renderItem = null;
+		resultScrollLeft = null;
+		resultScrollTop = null;
+		startScrollTop = null;
+		svgGenerator = null;
 	});
 
 	test('should render a list of \'items\'', () => {
@@ -64,7 +112,6 @@ describe('VirtualGridList with native \'scrollMode\'', () => {
 				dataSize={dataSize}
 				itemRenderer={renderItem}
 				itemSize={itemSize}
-				scrollMode="native"
 			/>
 		);
 
@@ -80,7 +127,6 @@ describe('VirtualGridList with native \'scrollMode\'', () => {
 				dataSize={dataSize}
 				itemRenderer={renderItem}
 				itemSize={itemSize}
-				scrollMode="native"
 			/>
 		);
 
@@ -97,7 +143,6 @@ describe('VirtualGridList with native \'scrollMode\'', () => {
 				dataSize={dataSize}
 				itemRenderer={renderItem}
 				itemSize={itemSize}
-				scrollMode="native"
 			/>
 		);
 
@@ -114,7 +159,6 @@ describe('VirtualGridList with native \'scrollMode\'', () => {
 				dataSize={dataSize}
 				itemRenderer={renderItem}
 				itemSize={itemSize}
-				scrollMode="native"
 			/>
 		);
 
@@ -126,7 +170,6 @@ describe('VirtualGridList with native \'scrollMode\'', () => {
 				dataSize={dataSize}
 				itemRenderer={renderItem}
 				itemSize={itemSize}
-				scrollMode="native"
 			/>
 		);
 
@@ -144,7 +187,6 @@ describe('VirtualGridList with native \'scrollMode\'', () => {
 				direction="horizontal"
 				itemRenderer={renderItem}
 				itemSize={itemSize}
-				scrollMode="native"
 			/>
 		);
 
@@ -164,7 +206,6 @@ describe('VirtualGridList with native \'scrollMode\'', () => {
 					itemRenderer={renderItem}
 					itemSize={itemSize}
 					horizontalScrollbar="visible"
-					scrollMode="native"
 					verticalScrollbar="visible"
 				/>
 			);
@@ -184,7 +225,6 @@ describe('VirtualGridList with native \'scrollMode\'', () => {
 					dataSize={dataSize}
 					itemRenderer={renderItem}
 					itemSize={itemSize}
-					scrollMode="native"
 				/>
 			);
 
@@ -206,7 +246,6 @@ describe('VirtualGridList with native \'scrollMode\'', () => {
 					itemRenderer={renderItem}
 					itemSize={itemSize}
 					horizontalScrollbar="hidden"
-					scrollMode="native"
 					verticalScrollbar="hidden"
 				/>
 			);
@@ -217,6 +256,325 @@ describe('VirtualGridList with native \'scrollMode\'', () => {
 
 			expect(verticalScrollbar).toBeNull();
 			expect(horizontalScrollbar).toBeNull();
+		});
+	});
+
+	describe('ScrollTo', () => {
+		test('should scroll to the specific item of a given \'index\' with scrollTo', (done) => {
+			const onScrollStop = handlerOnScrollStop(done, () => {
+				const expected = (240 + 16) * 2;
+
+				expect(resultScrollTop).toBe(expected);
+			});
+
+			render(
+				<VirtualGridList
+					cbScrollTo={getScrollTo}
+					clientSize={clientSize}
+					dataSize={dataSize}
+					itemRenderer={renderItem}
+					itemSize={itemSize}
+					onScrollStop={onScrollStop}
+				/>
+			);
+
+			act(() => myScrollTo({index: 9, animate: false}));
+		});
+
+		test('should scroll to the given \'x\' position with scrollTo', (done) => {
+			const onScrollStop = handlerOnScrollStop(done, () => {
+				const expected = 200;
+
+				expect(resultScrollLeft).toBe(expected);
+			});
+
+			render(
+				<VirtualGridList
+					cbScrollTo={getScrollTo}
+					clientSize={clientSize}
+					dataSize={dataSize}
+					direction="horizontal"
+					itemRenderer={renderItem}
+					itemSize={itemSize}
+					onScrollStop={onScrollStop}
+				/>
+			);
+
+			act(() => myScrollTo({position: {x: 200}, animate: false}));
+		});
+
+		test('should scroll to the given \'y\' position with scrollTo', (done) => {
+			const onScrollStop = handlerOnScrollStop(done, () => {
+				const expected = 200;
+
+				expect(resultScrollTop).toBe(expected);
+			});
+
+			render(
+				<VirtualGridList
+					cbScrollTo={getScrollTo}
+					clientSize={clientSize}
+					dataSize={dataSize}
+					itemRenderer={renderItem}
+					itemSize={itemSize}
+					onScrollStop={onScrollStop}
+				/>
+			);
+
+			act(() => myScrollTo({position: {y: 200}, animate: false}));
+		});
+
+		test('should scroll to the given \'align\' with scrollTo', (done) => {
+			const onScrollStop = handlerOnScrollStop(done, () => {
+				const expected = Math.ceil(dataSize / 4) * (itemSize.minHeight + 16) - clientSize.clientHeight;
+				const actual = resultScrollTop;
+
+				expect(actual).toBe(expected);
+			});
+
+			render(
+				<VirtualGridList
+					cbScrollTo={getScrollTo}
+					clientSize={clientSize}
+					dataSize={dataSize}
+					itemRenderer={renderItem}
+					itemSize={itemSize}
+					onScrollStop={onScrollStop}
+				/>
+			);
+
+			act(() => myScrollTo({align: 'bottom', animate: false}));
+		});
+
+		test('should scroll to the given \'align\' with scrollTo after changing \'dataSize\'', (done) => {
+			const newDataSize = 50;
+
+			const onScrollStop = handlerOnScrollStop(done, () => {
+				const expected = Math.ceil(newDataSize / 4) * (itemSize.minHeight + 16) - clientSize.clientHeight;
+				const actual = resultScrollTop;
+
+				expect(actual).toBe(expected);
+			});
+
+			const {rerender} = render(
+				<VirtualGridList
+					cbScrollTo={getScrollTo}
+					clientSize={clientSize}
+					dataSize={dataSize}
+					itemRenderer={renderItem}
+					itemSize={itemSize}
+					onScrollStop={onScrollStop}
+				/>
+			);
+
+			rerender(
+				<VirtualGridList
+					cbScrollTo={getScrollTo}
+					clientSize={clientSize}
+					dataSize={newDataSize}
+					itemRenderer={renderItem}
+					itemSize={itemSize}
+					onScrollStop={onScrollStop}
+				/>
+			);
+
+			act(() => myScrollTo({align: 'bottom', animate: false}));
+		});
+
+		test('should scroll to the given \'align\' with scrollTo after changing \'itemSize\'', (done) => {
+			const newItemSize = {minWidth: 300, minHeight: 270};
+
+			const onScrollStop = handlerOnScrollStop(done, () => {
+				const expected = Math.ceil(dataSize / 4) * (newItemSize.minHeight + 18) - clientSize.clientHeight;
+				const actual = resultScrollTop;
+
+				expect(actual).toBe(expected);
+			});
+
+			const {rerender} = render(
+				<VirtualGridList
+					cbScrollTo={getScrollTo}
+					clientSize={clientSize}
+					dataSize={dataSize}
+					itemRenderer={renderItem}
+					itemSize={itemSize}
+					onScrollStop={onScrollStop}
+				/>
+			);
+
+			rerender(
+				<VirtualGridList
+					cbScrollTo={getScrollTo}
+					clientSize={clientSize}
+					dataSize={dataSize}
+					itemRenderer={renderItem}
+					itemSize={newItemSize}
+					onScrollStop={onScrollStop}
+				/>
+			);
+
+			act(() => myScrollTo({align: 'bottom', animate: false}));
+		});
+
+		test('should scroll to the given \'align\' with scrollTo after changing \'spacing\'', (done) => {
+			const newSpacing = 6;
+
+			const onScrollStop = handlerOnScrollStop(done, () => {
+				const expected = Math.ceil(dataSize / 4) * (itemSize.minHeight + 16) - clientSize.clientHeight + 94; // Cannot calculate the exact length after applying spacing.
+				const actual = resultScrollTop;
+
+				expect(actual).toBe(expected);
+			});
+
+			const {rerender} = render(
+				<VirtualGridList
+					cbScrollTo={getScrollTo}
+					clientSize={clientSize}
+					dataSize={dataSize}
+					itemRenderer={renderItem}
+					itemSize={itemSize}
+					onScrollStop={onScrollStop}
+				/>
+			);
+
+			rerender(
+				<VirtualGridList
+					cbScrollTo={getScrollTo}
+					clientSize={clientSize}
+					dataSize={dataSize}
+					itemRenderer={renderItem}
+					itemSize={itemSize}
+					onScrollStop={onScrollStop}
+					spacing={newSpacing}
+				/>
+			);
+
+			act(() => myScrollTo({align: 'bottom', animate: false}));
+		});
+
+		describe('scroll events', () => {
+			test('should call onScrollStart once', () => {
+				render(
+					<VirtualGridList
+						cbScrollTo={getScrollTo}
+						clientSize={clientSize}
+						dataSize={dataSize}
+						itemRenderer={renderItem}
+						itemSize={itemSize}
+						onScrollStart={handlerOnScrollStart}
+					/>
+				);
+
+				act(() => myScrollTo({position: {y: 200}, animate: false}));
+
+				const expected = 1;
+
+				expect(onScrollStartCount).toBe(expected);
+			});
+
+			test('should call onScroll once', () => {
+				render(
+					<VirtualGridList
+						cbScrollTo={getScrollTo}
+						clientSize={clientSize}
+						dataSize={dataSize}
+						itemRenderer={renderItem}
+						itemSize={itemSize}
+						onScroll={handlerOnScroll}
+					/>
+				);
+
+				act(() => myScrollTo({position: {y: 200}, animate: false}));
+
+				const expected = 1;
+
+				expect(onScrollCount).toBe(expected);
+			});
+
+			test('should call onScrollStop once', (done) => {
+				const onScrollStop = handlerOnScrollStop(done, () => {
+					const expected = 1;
+
+					expect(onScrollStopCount).toBe(expected);
+				});
+
+				render(
+					<VirtualGridList
+						cbScrollTo={getScrollTo}
+						clientSize={clientSize}
+						dataSize={dataSize}
+						itemRenderer={renderItem}
+						itemSize={itemSize}
+						onScrollStop={onScrollStop}
+					/>
+				);
+
+				act(() => myScrollTo({position: {y: 200}, animate: false}));
+			});
+		});
+	});
+
+	describe('Scroll by event', () => {
+		beforeEach(() => {
+			jest.useFakeTimers();
+		});
+
+		afterEach(() => {
+			jest.useRealTimers();
+		});
+
+		test('should scroll by wheel', (done) => {
+			const fn = jest.fn();
+
+			const onScrollStop = handlerOnScrollStop(done, () => {
+				fn();
+				expect(startScrollTop).toBe(0);
+				expect(onScrollStartCount).toBe(1);
+				expect(resultScrollTop).toBeGreaterThan(0);
+			});
+
+			render(
+				<VirtualGridList
+					clientSize={clientSize}
+					dataSize={dataSize}
+					itemRenderer={renderItem}
+					itemSize={itemSize}
+					onScrollStop={onScrollStop}
+					onScrollStart={handlerOnScrollStart}
+				/>
+			);
+
+			const list = screen.getByRole('list');
+			fireEvent.wheel(list, {deltaY: 100});
+
+			act(() => jest.advanceTimersByTime(1000)); // Wait onScrollStop
+
+			expect(fn).toBeCalled();
+		});
+
+		test('should not scroll by wheel when \'noScrollByWheel\' prop is true', (done) => {
+			const fn = jest.fn();
+
+			render(
+				<VirtualGridList
+					clientSize={clientSize}
+					dataSize={dataSize}
+					itemRenderer={renderItem}
+					itemSize={itemSize}
+					noScrollByWheel
+					onScrollStop={fn}
+					onScrollStart={handlerOnScrollStart}
+				/>
+			);
+
+			const list = screen.getByRole('list');
+			fireEvent.wheel(list, {deltaY: 100});
+
+			act(() => jest.advanceTimersByTime(1000)); // Wait onScrollStop
+
+			expect(fn).not.toHaveBeenCalled();
+
+			done();
 		});
 	});
 
@@ -237,7 +595,6 @@ describe('VirtualGridList with native \'scrollMode\'', () => {
 					dataSize={itemArray.length}
 					itemRenderer={renderItemArray}
 					itemSize={itemSize}
-					scrollMode="native"
 				/>
 			);
 
@@ -247,7 +604,6 @@ describe('VirtualGridList with native \'scrollMode\'', () => {
 				dataSize={itemArray.length}
 				itemRenderer={renderItemArray}
 				itemSize={itemSize}
-				scrollMode="native"
 			/>);
 
 			jest.useFakeTimers();
